[build-utils] modules-build-helper: simplify process-catalog
by not generating both a DS file and a SPI wrapper Java class for each
script or data-type, but instead only a single Java class and leaving
the rest of the work up to ds-to-spi-maven-plugin and
maven-bundle-plugin.

Also make the _dsannotations instructions explicit everywhere. This
makes it immediately clear which services a bundle exposes, which is
especially useful for the generated classes. It is also needed in order
to not include services in the same packages in other bundles.

Also make the Private-Package instructions explicit everywhere. It is
needed to specify a "split-package" directive for some packages, and if
specified all private packages must be listed. For the sake of
simplicify and clarity we always include the instruction.

Also specify _dsannotations through a property "expose-services", which
can also be used by the ds-to-spi-maven-plugin.

Also fix "split-package" warnings.